feat(wallet): fractional quantities for treasury and crypto entries
Quantity becomes BigDecimal end to end. A new quantity_dec numeric
column stores the exact value (the legacy NOT NULL integer column keeps
a rounded copy; ddl-auto cannot retype it). Average price, realized
profit, sell validation, dividend sync, evolution and the finance
invested lookup all use the decimal quantity.

+++ profitly-api/src/main/java/tech/lougon/profitly/finance/infrastructure/wallet/WalletInvestedLookup.java
@@ -29,7 +29,7 @@ public class WalletInvestedLookup implements InvestedLookup {
                 .filter(entry -> entry.date() != null && !entry.date().isBefore(since))
                 .filter(entry -> entry.quantity() != null && entry.paidPrice() != null)
                 .filter(entry -> entry.typeOrBuy() == tech.lougon.profitly.wallet.domain.model.EntryType.BUY)
-                .map(entry -> entry.paidPrice().multiply(BigDecimal.valueOf(entry.quantity())))
+                .map(entry -> entry.paidPrice().multiply(entry.quantity()))
                 .reduce(BigDecimal.ZERO, BigDecimal::add);
     }
 }
